+++ src/plugins/switch/elan/qsw.h
@@ -110,7 +110,7 @@ int		qsw_pack_jobinfo(qsw_jobinfo_t j, Buf buffer);
 int		qsw_unpack_jobinfo(qsw_jobinfo_t j, Buf buffer);
 
 int 		qsw_setup_jobinfo(qsw_jobinfo_t j, int ntasks, 
-			bitstr_t *nodeset, uint32_t *tasks_per_node,
+			bitstr_t *nodeset, uint16_t *tasks_per_node,
 			int cyclic_alloc);
 void		qsw_teardown_jobinfo(qsw_jobinfo_t j);
