+++ src/plugins/switch/none/switch_none.c
@@ -118,7 +118,7 @@ int switch_p_alloc_jobinfo ( switch_jobinfo_t *switch_job )
 }
 
 int switch_p_build_jobinfo ( switch_jobinfo_t switch_job, char *nodelist, 
-		uint32_t *tasks_per_node, int cyclic_alloc, char *network)
+		uint16_t *tasks_per_node, int cyclic_alloc, char *network)
 {
 	return SLURM_SUCCESS;
 }
